[fix] CacheRequest - 직렬화 오류 - NoArgs 생성자 추가
