+++ api-client/cache-client/src/main/java/woowa/team4/bff/api/client/cache/request/CacheRequest.java
@@ -3,10 +3,14 @@ package woowa.team4.bff.api.client.cache.request;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 
+@ToString
 @Getter
 @Setter
+@NoArgsConstructor
 @AllArgsConstructor
 public class CacheRequest {
