+++ api-client/search-client/src/main/java/woowa/team4/bff/api/client/response/SearchResponse.java
@@ -1,0 +1,18 @@
+package woowa.team4.bff.api.client.response;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class SearchResponse {
+
+    List<Long> ids;
+}
