--- api-client/search-client/src/main/java/woowa/tream4/bff/api/client/response/SearchResponse.java
@@ -1,12 +1,0 @@
-package woowa.tream4.bff.api.client.response;
-
-import java.util.List;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SearchResponse {
-
-    List<Long> ids;
-}
