+++ domain/cache-domain-redis/src/main/java/woowa/team4/bff/cache/redis/service/CacheRedisService.java
@@ -1,7 +1,5 @@
 package woowa.team4.bff.cache.redis.service;
 
-import static com.fasterxml.jackson.databind.type.LogicalType.Map;
-
 import com.fasterxml.jackson.core.type.TypeReference;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -48,6 +46,7 @@ public class CacheRedisService implements CacheService {
         String json = valueOperations.get(restaurantKey);
         if (json == null) {
             RestaurantSummary restaurantSummary = restaurantReviewStatisticsRepository.findByRestaurantId(id);
+            // ToDo: Event Publisher
             valueOperations.set(restaurantKey, jsonConverter.convert(restaurantSummary));
             return restaurantSummary;
         }
