+++ domain/exposure-common-domain/src/main/java/woowa/team4/bff/domain/RestaurantSummary.java
@@ -5,7 +5,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
+@ToString
 @Getter
 @NoArgsConstructor
 @Builder
