+++ domain/exposure-domain/build.gradle
@@ -6,14 +6,11 @@ dependencies {
     // module
     implementation project(":core-web")
     implementation project(":api-client:cache-client")
+    implementation project(":api-client:search-client")
     implementation project(":api-client:delivery-client")
     implementation project(":api-client:coupon-client")
     implementation project(":api-client:advertisement-client")
     implementation project(":domain:exposure-common-domain")
     implementation project(":domain:search-domain-rdb")
     // implementation project(":domain:search-domain-es")
-    implementation project(":domain:cache-domain-redis")
-    implementation project(":event:event-publisher")
-    implementation project(":event:common-event")
-    implementation project(":event:cache-event")
 }
