+++ domain/exposure-domain/src/main/java/woowa/team4/bff/exposure/external/caller/SyncExternalApiCaller.java
@@ -32,4 +32,5 @@ public class SyncExternalApiCaller {
     public List<AdvertisementResponse> getAdvertisement(List<Long> restaurantIds, String keyword) {
         return advertisementApiCaller.send(new AdvertisementRequest(restaurantIds, keyword));
     }
+
 }
