+++ domain/search-domain-rdb/build.gradle
@@ -1,5 +1,7 @@
 dependencies {
     // module
     implementation project(":domain:restaurant-domain-rdb")
+    implementation project(":domain:menu-domain-rdb")
+    implementation project(":domain:review-domain-rdb")
     implementation project(":domain:exposure-common-domain")
 }
