--- mono/build.gradle
@@ -1,60 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.3.2'
-    id 'io.spring.dependency-management' version '1.1.6'
-}
-
-group = 'woowa.team4'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-ext {
-    set('springCloudVersion', "2023.0.3")
-}
-
-
-bootJar.enabled = true
-jar.enabled = false
-
-dependencies {
-    // web
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    // aop
-    implementation 'org.springframework.boot:spring-boot-starter-aop'
-    // util
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    // module
-    implementation project(":core-web")
-    implementation project(":domain:menu-domain-rdb")
-    implementation project(":domain:restaurant-domain-rdb")
-    implementation project(":domain:review-domain-rdb")
-    //
-    implementation project(":domain:exposure-common-domain")
-    implementation project(":domain:exposure-domain")
-}
-
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
