--- mono/src/main/java/woowa/team4/bff/BffApplication.java
@@ -1,15 +1,0 @@
-package woowa.team4.bff;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
-
-@EnableAsync
-@SpringBootApplication
-public class BffApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BffApplication.class, args);
-    }
-
-}
