--- mono/src/main/resources/application.yml
@@ -1,90 +1,0 @@
-spring:
-  profiles:
-    active: local
-
----
-spring:
-  config:
-    activate:
-      on-profile: local
-  datasource:
-    url: jdbc:mysql://localhost:3306/mydb
-    username: user
-    password: userpassword
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    database-platform: org.hibernate.dialect.MySQLDialect
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-    hibernate:
-      ddl-auto: none
-
-  elasticsearch:
-    uris: http://localhost:9200
-    username: elastic
-    password: your_elastic_password
-  data:
-    elasticsearch:
-      repositories:
-        enabled: true
-    redis:
-      host: localhost
-      port: 6379
-
-    mongo:
-      host: localhost
-      port: 27000
-      database: "mydb"
-      username: "user"
-      password: "userpassword"
-
-logging:
-  level:
-    tracer: TRACE
-
-external-api:
-  advertisement-service:
-    endpoints: http://localhost:8083/api/v1/advertisements
-  coupon-service:
-    endpoints: http://localhost:8082/api/v1/coupons
-  delivery-service:
-    endpoints: http://localhost:8081/api/v1/deliveries
-
----
-spring:
-  config:
-    activate:
-      on-profile: dev
-  datasource:
-    url: jdbc:mysql://${DB_URL}
-    username: ${DB_USER}
-    password: ${DB_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    database-platform: org.hibernate.dialect.MySQLDialect
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-    hibernate:
-      ddl-auto: none
----
-spring:
-  config:
-    activate:
-      on-profile: prod
-  datasource:
-    url: jdbc:mysql://${DB_URL}
-    username: ${DB_USER}
-    password: ${DB_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    database-platform: org.hibernate.dialect.MySQLDialect
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-    hibernate:
-      ddl-auto: none
