+++ service/advertisement-service/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 8083
+  port: 8081
 
 spring:
   threads:
