+++ service/cache-service/src/main/java/woowa/team4/bff/cacheservice/CacheController.java
@@ -2,13 +2,17 @@ package woowa.team4.bff.cacheservice;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import woowa.team4.bff.api.client.cache.request.CacheRequest;
 import woowa.team4.bff.api.client.cache.response.CacheResponse;
+import woowa.team4.bff.domain.RestaurantSummary;
 import woowa.team4.bff.interfaces.CacheService;
 
+@Slf4j
 @RequestMapping("/api/v1/cache")
 @RestController
 @RequiredArgsConstructor
@@ -16,9 +20,12 @@ public class CacheController {
 
     private final CacheService cacheService;
 
-    @GetMapping("")
-    public List<CacheResponse> findByRestaurantIds(CacheRequest request){
-        return cacheService.findByRestaurantIds(request.getIds()).stream()
+    @PostMapping("")
+    public List<CacheResponse> findByRestaurantIds(@RequestBody CacheRequest request){
+        log.info("[findByRestaurantIds] request: {}", request);
+        List<RestaurantSummary> res = cacheService.findByRestaurantIds(request.getIds());
+        log.info("[findByRestaurantIds] response: {}", res);
+        return res.stream()
                 .map(e -> CacheResponse.builder()
                         .restaurantId(e.getId())
                         .restaurantUuid(e.getRestaurantUuid())
