+++ service/coupon-service/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 8082
+  port: 8083
 
 spring:
   threads:
