+++ service/delivery-time-service/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 8081
+  port: 8084
 
 spring:
   threads:
