--- service/restaurant-exposure-service/src/test/java/woowa/team4/bff/RestauarntExposureServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.bff;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestauarntExposureServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
